Optimize ranked guild calculations in rank variable resolver
We're currently creating a collection of ranking guilds based on
the number of members in the guild each time the rank variable resolver
is being used.

This change will speed up the rank variable resolver,
because we won't be calculating the entire collection on every use -
instead we will be retrieving following guilds until they will
meet the given criteria.