GH-1159 Return "dummy" user objects when uuid is version 2
We want not to lookup/create user object for player with version 2 uuid, as players with this version are probably a NPC. We want to return a "dummy" object, because currently with our database system handling these NPC's as valid users would lead to save these "invalid" (as NPC isn't a real player, so isn't a valid user) users into database, and that's not what we want.

This commit also updates GH-1155.