GH-1410 Handle guild terrain explosions properly
Before, when TNT (or other explosive) exploded on terrain that was not a guild terrain, but the explosion reached the blocks that were on the guild terrain, these blocks would explode even if this guild should be protected from explosions at the time.

Now, we're excluding blocks from the explosion that belong to the guild's terrain that have explosion protection even if explosion source was not located on the guild's terrain.

Closes #1410