Fix interaction between our plugin and others while handling explosion event
Previously, we've been handling explosion event improperly, because we've not been taking into an account the fact that other plugins (for example WorldGuard) often doesn't cancel the event, but instead they are clearing the list of the blocks that have been affected by the explosion and because of that our custom block explosion logic could be fired even if it shouldn't be.

Now, we're calculating all the blocks that we need for our purposes before most of the protection (but not only) plugins code is fired and modifying the explosion event block list so that blocks that we're interested in are included, then let other plugins run their event handlers, and then based on the resulting event data we can handle explosion event properly without breaking interaction with other plugins.