GH-1594 Set guild build block if explosion event hasn't been cancelled
Before, we've been setting guild build block even if later we would've
cancelled the explosion because of various TNT explosion control options.

Now, we will be trying to set guild build block at the end of the
event handler chain, so that if the explosion event will be cancelled
(by us or someone else), we will properly do not set the guild build
block.