Do not send player list header and footer if they are both empty
Before migration to modules, player list implementation wouldn't send
header and footer if they were an empty strings.

Current behavior causes flickering when there is something else on the
server that is sending header/footer packets and our player list is in use.

Thus, send player list header and footer only if they are defined to be empty
to prevent this issue from occurring.