GH-2516 Rework NMS modules and fix 1.20.4 player list (#2544)
Rework NMS modules to use paperweight-userdev wherever possible to avoid further issues with wrong NMS objects being used because of missing mappings. This also fixes #2516 with player list not being displayed properly on 1.20.4.

This PR also removes support for older revisions for versions < 1.19.3, as from now our policy for supporting versions is to support all revisions only for the latest version of the game.